requirements.txt
openai==0.28.0
python-dotenv==1.1.1
streamlit==1.47.0
